Various adjustments per comments on GitHub
