+++ rcjaRegistration/events/admin.py
@@ -27,14 +27,14 @@ from regions.admin import StateAdmin
 class DivisionCategoryAdmin(AdminPermissions, admin.ModelAdmin):
     pass
 
-class ActiveFilter(admin.SimpleListFilter):
+class DivisionActiveFilter(admin.SimpleListFilter):
     title = _('Active')
 
     parameter_name = 'active'
 
     def lookups(self, request, model_admin):
         return (
-            ('active', _('Active')),
+            (None, _('Active')),
             ('inactive', _('Inactive')),
             ('all', _('All'))
         )
@@ -185,7 +185,7 @@ class AvailableDivisionInline(FKActionsRemove, InlineAdminPermissions, admin.Tab
     def fkObjectFilterFields(cls, request, obj):
         return {
             'division': {
-                'queryset': Division.objects.filter((Q(state=obj.state) | Q(state=None))& Q(active=True)) if obj is not None else Division.objects.none(), # Inline not displayed on create so user will never see fallback to None
+                'queryset': Division.objects.filter((Q(state=obj.state) | Q(state=None))) if obj is not None else Division.objects.none(), # Inline not displayed on create so user will never see fallback to None
             },
         }
